+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# 1.7.6
+# 1.8.0 – Config Changes
+- Move worldgeneration config to JSON from json5 to match all other configs
+- Correctly update missing keys in JSON configs
+- Automatically migrate legacy config files to the new format (Will be removed in 1.9.0)
+- Add more comments to better clarify things to users in config files
 
 # 1.7.5
 - Fix Incorrect English Language key for Granny Smith Advancement